requirements.txt
alembic==1.0.11
astroid==2.2.5
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
Mako==1.0.12
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.3
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
python-decouple==3.1
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.5
typed-ast==1.4.0
Werkzeug==0.15.4
wrapt==1.11.2
